pyproject.toml
[tool.poetry]
name = "twillio-secret-santa"
version = "1.0.0"
description = "A python script that generates secret santa allocations given some restraints and sends the allocations to participants using Twilio."
authors = ["Daniel Kirkwood <danielkirkwood1973@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
twilio = "^9.3.3"
networkx = "^3.4.1"
python-dotenv = "^1.0.1"
matplotlib = "^3.9.2"
aiosqlite = "^0.20.0"
tortoise-orm = {extras = ["aiosqlite"], version = "^0.21.7"}
python-constraint = "^1.4.0"
click = "^8.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
